kinah_website_backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.6
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
frozenlist==1.8.0
gunicorn==25.1.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
multidict==6.7.1
mysqlclient==2.2.8
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.3
twilio==9.10.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
yarl==1.23.0
